added input system

+++ lib/MovementSystem.hpp
@@ -7,5 +7,5 @@ class GameState;
 
 class MovementSystem {
 public:
-  void Update(GameState &state, uint64_t deltatime);
+  void Update(GameState &state, float deltatime);
 };
